refactor: wrap quantity
- 수량 원시값 객체 포장
- jpa converter 추가
- converter auto apply true 추가
- 패키지 분리 infra -> generic

+++ src/main/java/kitchenpos/order/dto/OrderLineRequest.java
@@ -2,12 +2,12 @@ package kitchenpos.order.dto;
 
 public class OrderLineRequest {
     private long menuId;
-    private int quantity;
+    private long quantity;
 
     public OrderLineRequest() {
     }
 
-    public OrderLineRequest(long menuId, int quantity) {
+    public OrderLineRequest(long menuId, long quantity) {
         this.menuId = menuId;
         this.quantity = quantity;
     }
@@ -16,7 +16,7 @@ public class OrderLineRequest {
         return menuId;
     }
 
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
 }
